Remove indirection via Client (Command->Client->Operation) when calling Operations; reduced to Command->Operation
