+++ syncany-cli/src/main/java/org/syncany/cli/GenlinkCommand.java
@@ -23,6 +23,7 @@ import joptsimple.OptionSet;
 import joptsimple.OptionSpec;
 
 import org.syncany.operations.OperationResult;
+import org.syncany.operations.init.GenlinkOperation;
 import org.syncany.operations.init.GenlinkOperationOptions;
 import org.syncany.operations.init.GenlinkOperationResult;
 
@@ -42,7 +43,7 @@ public class GenlinkCommand extends AbstractInitCommand {
 	@Override
 	public int execute(String[] operationArgs) throws Exception {
 		GenlinkOperationOptions operationOptions = parseOptions(operationArgs);		
-		GenlinkOperationResult operationResult = client.genlink(operationOptions);		
+		GenlinkOperationResult operationResult = new GenlinkOperation(config, operationOptions).execute();		
 		
 		printResults(operationResult);
